requirements.txt
Kivy~=2.0.0
kivymd~=0.104.2.dev0
uvicorn~=0.13.4
fastapi~=0.65.1
wolframclient~=1.1.4
opencv-python~=4.5.1.48
numpy~=1.19.5
matplotlib~=3.4.1
tensorflow-cpu==2.5.0
python-multipart